Add files via upload
A fitness center has a high demand for bikes from 6:00-7:30 am. Since we want to model the arrival of the clients, to simplify the problem, we assume they only arrive at discrete time intervals. For example, no client would arrive 30 seconds past the 5th minute; they only arrive at 5th or 6th minute. The assumptions are: 
1-	First come first serve basis
2-	Clients wait in the line when there are at most 3 people ahead of them
3-	Clients who are in the line already, will not leave until they get a bike
4-	Clients use the bikes for exactly 8 minutes each 
In order to make the model even simpler, we assume that all clients use bikes for exactly 8 minutes. Otherwise, we would have a complicated availability for each bike. The rates of arrivals are as follows: 6:00-6:30: 30 people, 6:31-7:00 45 people, 7:01-7:30 20 people. People who arrive when all bikes have 3 people in the line, will leave the gym unsatisfied. If a client becomes dissatisfied for more than 4 times in a month, there is 50% chance that they drop their membership. A complete process of this model is shown in the flowchart in FlowchartUnit3.pdf.